requirements.txt
flask==3.0.2
psycopg2-binary==2.9.9
polygon-api-client==1.12.3
python-dotenv==1.0.1
pandas==2.2.1
SQLAlchemy==2.0.27
plotly==5.19.0
dash==2.15.0
requests==2.31.0 